Fix #9772: Deprecate widgetBuilder init method
