+++ primefaces-showcase/src/main/java/org/primefaces/showcase/view/panel/DashboardView.java
@@ -58,9 +58,9 @@ public class DashboardView implements Serializable {
 
         // legacy
         legacyModel = new DefaultDashboardModel();
-        legacyModel.addWidget(new DefaultDashboardColumn(Arrays.asList("sports", "finance")));
-        legacyModel.addWidget(new DefaultDashboardColumn(Arrays.asList("lifestyle", "weather")));
-        legacyModel.addWidget(new DefaultDashboardColumn(Arrays.asList("politics")));
+        legacyModel.addWidget(new DefaultDashboardWidget(Arrays.asList("sports", "finance")));
+        legacyModel.addWidget(new DefaultDashboardWidget(Arrays.asList("lifestyle", "weather")));
+        legacyModel.addWidget(new DefaultDashboardWidget(Arrays.asList("politics")));
     }
 
     public void handleReorder(DashboardReorderEvent event) {
