+++ primefaces/src/main/java/org/primefaces/model/dashboard/DefaultDashboardColumn.java
@@ -23,101 +23,33 @@
  */
 package org.primefaces.model.dashboard;
 
-import java.io.Serializable;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.LinkedList;
-import java.util.List;
 
 /**
  * @deprecated use DefaultDashboardWidget
  */
 @Deprecated
-public class DefaultDashboardColumn implements DashboardWidget, Serializable {
+public class DefaultDashboardColumn extends DefaultDashboardWidget {
 
     private static final long serialVersionUID = 1L;
 
-    private List<String> widgets;
-    private String style;
-    private String styleClass;
-
     public DefaultDashboardColumn() {
-        widgets = new LinkedList<>();
+        super();
     }
 
     public DefaultDashboardColumn(String widgetId, String styleClass) {
-        this();
+        super(widgetId, styleClass);
         getWidgets().addAll(Arrays.asList(widgetId));
         setStyleClass(styleClass);
     }
 
     public DefaultDashboardColumn(String style, String styleClass, Collection<String> widgets) {
-        this();
-        this.widgets.addAll(widgets);
-        this.style = style;
-        this.styleClass = styleClass;
+        super(style, styleClass, widgets);
     }
 
     public DefaultDashboardColumn(Collection<String> widgets) {
-        this();
-        this.widgets.addAll(widgets);
+        super(widgets);
     }
-
-    @Override
-    public void removeWidget(String widgetId) {
-        widgets.remove(widgetId);
-    }
-
-    @Override
-    public List<String> getWidgets() {
-        return widgets;
-    }
-
-    @Override
-    public int getWidgetCount() {
-        return widgets.size();
-    }
-
-    @Override
-    public String getWidget(int index) {
-        return widgets.get(index);
-    }
-
-    @Override
-    public void addWidget(int index, String widgetId) {
-        widgets.add(index, widgetId);
-    }
-
-    @Override
-    public void reorderWidget(int index, String widgetId) {
-        widgets.remove(widgetId);
-        widgets.add(index, widgetId);
-    }
-
-    @Override
-    public void addWidget(String widgetId) {
-        widgets.add(widgetId);
-    }
-
-    @Override
-    public String getStyle() {
-        return style;
-    }
-
-    @Override
-    public void setStyle(String style) {
-        this.style = style;
-    }
-
-    @Override
-    public String getStyleClass() {
-        return styleClass;
-    }
-
-    @Override
-    public void setStyleClass(String styleClass) {
-        this.styleClass = styleClass;
-    }
-
 
 }
